[CB-230] add dialog when trying to resign and delete unnecessary buttons in login page
